refactor: vertical splitted list to unite
